added "async" to lines 59 and 67
included in a later part of the tutorial (line 95), a line of code reads: "client.on(Events.InteractionCreate, async interaction => {" and it's at a part where it expects you to have already written it down by that point. However the line includes "async", which was not written in the previous example sections on lines 59 and 67. This change is to fix that.